Color Bitbucket Cloud header
This also moves the API queries to a separate module to respect
Rubocop's class length limit.

Resolves #276
