+++ lib/repofetch/bitbucketcloud.rb
@@ -3,6 +3,7 @@
 require 'action_view'
 require 'optparse'
 require 'repofetch'
+require 'repofetch/bitbucketcloud/stats'
 require 'repofetch/exceptions'
 require 'sawyer'
 
@@ -10,6 +11,7 @@ class Repofetch
   # Adds support for Bitbucket repositories.
   class BitbucketCloud < Repofetch::Plugin
     include ActionView::Helpers::NumberHelper
+    include Repofetch::BitbucketCloud::Stats
 
     HTTP_REMOTE_REGEX = %r{https?://bitbucket\.org/(?<owner>[\w._-]+)/(?<repo>[\w._-]+)}.freeze
     SSH_REMOTE_REGEX = %r{git@bitbucket\.org:(?<owner>[\w._-]+)/(?<repo>[\w._-]+)}.freeze
@@ -24,7 +26,13 @@ class Repofetch
     end
 
     def header
-      "#{repo_data['owner']['display_name']}/#{repo_data['name']} @ Bitbucket"
+      name = header_format("#{repo_data['owner']['display_name']}/#{repo_data['name']}")
+      host = header_format('Bitbucket')
+      "#{name} @ #{host}"
+    end
+
+    def header_format(text)
+      theme.format(:bold, theme.format(:blue, text))
     end
 
     def stats
@@ -97,60 +105,6 @@ class Repofetch
       raise Repofetch::PluginUsageError, parser.to_s unless args.length == 1
 
       new(args[0])
-    end
-
-    protected
-
-    def repo_data
-      @repo_data ||= agent.call(:get, "repositories/#{@repo_identifier}").data
-    end
-
-    def clone_urls
-      @clone_urls ||= repo_data['links']['clone'].to_h { |clone| [clone['name'].to_sym, clone['href']] }
-    end
-
-    def http_clone_url
-      Repofetch::Stat.new('HTTP(S)', clone_urls[:https], emoji: '🌐')
-    end
-
-    def ssh_clone_url
-      Repofetch::Stat.new('SSH', clone_urls[:ssh], emoji: '🔑')
-    end
-
-    def watchers
-      @watcher_data ||= agent.call(:get, "repositories/#{@repo_identifier}/watchers").data
-      Repofetch::Stat.new('subscribers', @watcher_data['size'], emoji: '👀')
-    end
-
-    def forks
-      @fork_data ||= agent.call(:get, "repositories/#{@repo_identifier}/forks").data
-      Repofetch::Stat.new('forks', @fork_data['size'], emoji: '🔱')
-    end
-
-    def created
-      Repofetch::TimespanStat.new('created', repo_data['created_on'], emoji: '🐣')
-    end
-
-    def updated
-      Repofetch::TimespanStat.new('updated', repo_data['updated_on'], emoji: '📤')
-    end
-
-    def size
-      # NOTE: Size is in bytes
-      # TODO: Move this somewhere else instead of using a copy-paste
-      byte_size = number_to_human_size(repo_data['size'] || 0, precision: 2, significant: false,
-                                                               strip_insignificant_zeros: false)
-      Repofetch::Stat.new('size', byte_size, emoji: '💽')
-    end
-
-    def issues
-      @issue_data ||= agent.call(:get, "repositories/#{@repo_identifier}/issues").data
-      Repofetch::Stat.new('issues', @issue_data['size'], emoji: '❗')
-    end
-
-    def pull_requests
-      @pull_request_data ||= agent.call(:get, "repositories/#{@repo_identifier}/pullrequests").data
-      Repofetch::Stat.new('pull requests', @pull_request_data['size'], emoji: '🔀')
     end
   end
 end
